+++ classes/AuthForm.php
@@ -5,8 +5,6 @@
 class AuthForm extends BaseForm {
     public $formName = 'auth';
 
-    public $user;
-
     protected $fields = ['email', 'password'];
 
     protected $rules = [
@@ -21,23 +19,19 @@ class AuthForm extends BaseForm {
     protected function runEmailValidator($fields)
     {
         $result = true;
-        $user = User::getUser($this->formData['email']);
+
         foreach($fields as $value){
             $field = $this->formData[$value];
-            if(filter_var($field, FILTER_VALIDATE_EMAIL) && !$user){
-                $result = false;
-                $this->errors[$value] = 'Пользователя с таким именем не существует, воспользуйтесь формой регистрации';
-            }else{
-                $this -> user = $user;
-            }
-            if(!filter_var($field, FILTER_VALIDATE_EMAIL)){
+            if (!filter_var($field, FILTER_VALIDATE_EMAIL)) {
                 $result = false;
                 $this->errors[$value] = 'Введите корректный e-mail';
             }
-
         }
-
         return $result;
+    }
+
+    public function addBadEmailOrPasswordError() {
+        $this->errors['email'] = 'Пользователя с таким именем  паролем не существует, воспользуйтесь формой регистрации';
     }
 
 }
