+++ classes/Database.php
@@ -6,17 +6,17 @@
 class Database
 {
     /**
-     * @var подключение к базе дынных
+     * @var mysqli подключение к базе дынных
      */
     private $link;
     /**
-     * @var содержит ошибку подключения к б/д
+     * @var string содержит ошибку подключения к б/д
      */
     private $error;
 
     /**
      * Фукция создает единственный объект подключения к базе данных
-     * @return object
+     * @return Database
      */
     public static function instance()
     {
@@ -33,20 +33,14 @@ class Database
      */
     private function __construct()
     {
-        include 'config.php';
-        $this->link = mysqli_connect($DBCONFIG['host'], $DBCONFIG['user'], $DBCONFIG['password'], $DBCONFIG['database']);
+        include dirname(__FILE__).'/../config.php';
+        /** @var array $DBCONFIG */
+        $this->link = @mysqli_connect($DBCONFIG['host'], $DBCONFIG['user'], $DBCONFIG['password'], $DBCONFIG['database']);
         if (!$this->link) {
-            $this->error = 'Ошибка: Невозможно подключиться к MySQL ' . mysqli_connect_error();
+            $this->error = 'Ошибка: Невозможно подключиться к MySQL. ' . mysqli_connect_error();
+            echo $this->error;
+            exit;
         }
-    }
-
-    /**
-     * Функция возвращает соединение
-     * @return array
-     */
-    public function resultConnection()
-    {
-        return $this->link;
     }
 
     /**
@@ -76,8 +70,8 @@ class Database
      */
     public function updateData($nameTable, $updateData = [], $where = [])
     {
-        $setPoints = convertAssocDataToWhereStmt($updateData);
-        $condition = convertAssocDataToWhereStmt($where);
+        $setPoints = Database::convertAssocDataToWhereStmt($updateData);
+        $condition = Database::convertAssocDataToWhereStmt($where);
 
         $sql = 'UPDATE ' . $nameTable . ' SET ' . $setPoints[0] . ' WHERE ' . $condition[0];
         $value = array_merge($setPoints[1], $condition[1]);
@@ -111,7 +105,7 @@ class Database
      * Функция получает данные из базы
      * @param string $sql - sql запрос
      * @param array $data массив с данными для запроса
-     * @return array $theResult  пустой массив или
+     * @return array $theResult  пустой массив или  // или что?
      */
     public function getData($sql, $data = [])
     {
@@ -137,7 +131,7 @@ class Database
      * Функция удаляет данные из базы
      * @param string $sql - sql запрос
      * @param array $data массив с данными для запроса
-     * @return array $theResult  пустой массив или
+     * @return int|bool
      */
     public function deleteData($sql, $data = [])
     {
@@ -145,8 +139,27 @@ class Database
         if ($stmt) {
             mysqli_stmt_execute($stmt);
             mysqli_stmt_get_result($stmt);
+            mysqli_stmt_close($stmt);
             return mysqli_insert_id($this->link);
         }
         return false;
+    }
+
+    /**
+     * Функция разбивает массив на 2 по ключу и значению
+     * @param  array $array массив для преобразования
+     * @return array , значениями которого я вляются 2 массива,
+     * 1 - это строка с плайсхолдерами, 2 строка со значениями для них
+     */
+    public static function convertAssocDataToWhereStmt($array = [])
+    {
+        if (!$array) {
+            return [];
+        }
+        $result = array_map(function ($k) {
+            return $k . ' ?';
+        }, array_keys($array));
+
+        return [implode(", ", $result), array_values($array)];
     }
 }
