+++ classes/User.php
@@ -22,7 +22,7 @@ class User
      * Функция добавляет пользователей в базу
      * @param  array $fields валидные и не валидные поля
      */
-    public static function addUserToDatabase($fields)
+    public static function addUser($fields)
     {
         $sqlAddUser = "INSERT INTO user(registered, email, name, password) VALUES (NOW(), ?, ?, ?)";
         Database::instance()->setData($sqlAddUser, [
@@ -33,13 +33,14 @@ class User
 
     /**
      * Функция находит пользователя по email в базе данных
-     * @param string email электронная почта
+     * @param string $email электронная почта
      * @return array $user если пользоваетель существует или null
      */
     public static function getUser($email)
     {
         $sql = "SELECT id, email, password, name FROM user WHERE email = ?";
-        return Database::instance()->getData($sql, [$email]);
+        $res = Database::instance()->getData($sql, [$email]);
+        return count($res) > 0 ? $res[0] : null;
     }
 
 }
