+++ logout.php
@@ -1,5 +1,5 @@
 <?php
 session_start();
-unset($_SESSION['user']);
+Auth::logout();
 header("Location: /index.php");
-exit();
+
