+++ register.php
@@ -5,13 +5,13 @@ $form = new SignupForm();
 if ($form->isSubmitted()) {
     $form->validate();
     if ($form->isValid()) {
-        User::addUserToDatabase($form->getformData());
+        User::addUser($form->getFormData());
         header("Location: /login.php?login=1&show_message=1");
     }
 }
 printHead();
-print(includeTemplate('header.php', ['user' => Auth::logout()]));
+print(includeTemplate('header.php', ['user' => Auth::getAuthUser()]));
 print(includeTemplate('register.php', ['form' => $form]));
 printEndDivLayout();
-print includeTemplate('footer.php', ['user' => Auth::logout()]);
+print includeTemplate('footer.php', ['user' => Auth::getAuthUser()]);
 printEndBodyHtml();
