+++ templates/main.php
@@ -30,14 +30,14 @@
             </ul>
         </nav>
 
-        <a class="button button--transparent button--plus content__side-button" href="/index.php?addCategory">Добавить
+        <a class="button button--transparent button--plus content__side-button" href="/index.php?addProject">Добавить
             проект</a>
     </section>
 
     <main class="content__main">
         <h2 class="content__main-heading">Список задач</h2>
 
-        <form class="search-form" action="index.php?search" method="get">
+        <form class="search-form" action="/index.php?search" method="get">
             <input class="search-form__input" type="text" name="search" value="" placeholder="Поиск по задачам">
 
             <input class="search-form__submit" type="submit" name="" value="Искать">
